Add Corteza OIDC support
This provides `well know` endpoint for corteza which defines a set of values that can be retrieved by a client in order to self-configure.

Added route for `public-keys` which can help a client to determine if the public key is changed since the last time and automatically update itself to use this new information.