Fix report request crash when using dup. attr idents
When request to /report endpoint with multiple attributes using the same
ident was used, request crashed with an error that Scan expects N+M
destinations but got N.

This was due to the way scan destination slice was constructed, using
column ident as unique key. Only problem for aggregations.

Example:
/report?metrics=SUM(Amount)+AS+sum_amount,SUM(Amount)+AS+sum_amount

Error:
sql: expected 4 destination arguments in Scan, not 3"

This is now done only for attributs with store type SimpleJsonDocColumn.

Ref: #465