Fix failed tests
This commit fixes a failure of `TestModuleUpdateWithReservedFieldName`

The fix is actually only indirectly related to the feature of allowing
adding of physical columns. If a system field already exists in a module
when doing an update, there is a promise of backwards compatibility
that intentionally ignored the field. This not taken into account
in the `ModuleToModel` function since it was adding system fields
without checking if the user had already added that field.