Replace defer with explicit mutex unlocks
Using defer to unlock the muytex makes it so only one operation can run at any point in time.
This is because the function passed in is run inside `withModel`.

Since we only need to protect the model cache, and not the operation themselves,
this is a case where its better to unlock the mutex at each code branch.